+++ vars/puppet.groovy
@@ -178,3 +178,13 @@ def task(body) {
   }
   print 'Puppet Orchestrator Task execution successfully requested.'
 }
+
+def token (body) {
+  // evaluate the body block and collect configuration into the object
+  def config = [:]
+  body.resolveStrategy = Closure.DELEGATE_FIRST
+  body.delegate = config
+  body()
+
+  // input checking
+}